Move Badges to its own file
