+++ src/components/organisms/Badges/index.ts
@@ -1,0 +1,1 @@
+export { Badges } from "./Badges";
